pagination and network errors handling

+++ app/src/main/java/com/example/quotes/ui/fragments/FavoritesFragment.kt
@@ -1,5 +1,7 @@
 package com.example.quotes.ui.fragments
 
+import android.app.AlertDialog
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -34,10 +36,29 @@ class FavoritesFragment : Fragment(R.layout.fragment_favorites) {
 
         adapter = SavedQuotesAdapter(object : QuoteClickListener {
             override fun likeOrUnLike(quote: QuoteUiModel) {
-                viewModel.deleteQuote(quote)
+                val builder = AlertDialog.Builder(context)
+                builder.setTitle("Delete")
+                builder.setMessage("Delete from your favorites?")
+                builder.setPositiveButton("Yes") { dialogInterface, _ ->
+                    viewModel.deleteQuote(quote)
+                    dialogInterface.dismiss()
+                }
+                builder.setNegativeButton("No") { dialogInterface, _ ->
+                    dialogInterface.dismiss()
+                }
+                builder.show()
             }
 
             override fun share(quote: QuoteUiModel) {
+                val text = "“${quote.content}”\n-${quote.author}"
+                val link = "Developer contact:\nhttps://t.me/elnurIsaevBlog"
+                val sendIntent = Intent().apply {
+                    action = Intent.ACTION_SEND
+                    putExtra(Intent.EXTRA_TEXT, "$text\n\n$link")
+                    type = "text/plain"
+                }
+                val shareIntent = Intent.createChooser(sendIntent, null)
+                startActivity(shareIntent)
             }
         })
         binding.favRV.adapter = adapter
